split build_energy_totals rule into energy and co2
